realtime request tests
