+++ test/src/v2/departures/index.ts
@@ -1,10 +1,9 @@
-export { realtime } from './departures';
+export { realtimeScenario, realtimeForQuayDepartures } from './realtime';
 export { stopDepartures } from './departures';
 export { stopDeparturesPOSTandGET } from './departures';
 export { quayDeparturesVsStopDepartures } from './departures';
 export { quayDepartures } from './departures';
 export { quayDeparturesPOSTandGET } from './departures';
-export { realtimeForQuayDepartures } from './departures';
 export { stopsNearest } from './stops';
 export { stopsDetails } from './stops';
 export { departureFavorites } from './favorites';
